*refactoring of ConvertSliceNode.
