+++ prj/vs2022/CvtOnnx.vcxproj.filters
@@ -273,6 +273,9 @@
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertSinNode.cpp">
       <Filter>OnnxRuntime\S</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertSliceNode.cpp">
+      <Filter>OnnxRuntime\S</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertSqueezeNode.cpp">
       <Filter>OnnxRuntime\S</Filter>
     </ClCompile>
